chore: Move backup logic
