+++ mobile/src/navigation/screens/settings/View/Sheets.tsx
@@ -3,7 +3,7 @@ import {
   useUserPreferencesStore,
 } from "~/services/UserPreferences";
 import { LANGUAGES } from "~/modules/i18n/constants";
-import { useExportBackup, useImportBackup } from "./Sheets.utils";
+import { useExportBackup, useImportBackup } from "~/modules/backup/JSON";
 
 import { mutateGuard } from "~/lib/react-query";
 import { FlatList, useIsScrollable } from "~/components/Defaults";
